requirements.txt
pypdf
google-genai
pandas
openpyxl
python-dotenv
tqdm
numpy
